Update TestCombineTuples to use int64 for expected field values
This commit modifies the TestCombineTuples function in the tuple_test.go file to expect int64 values instead of int32. This change aligns with recent updates to integer handling across the codebase, ensuring consistency in type usage during tests.

+++ pkg/tuple/tuple_test.go
@@ -320,7 +320,7 @@ func TestCombineTuples(t *testing.T) {
 					field, _ := result.GetField(i)
 
 					switch v := expected.(type) {
-					case int32:
+					case int64:
 						intField, ok := field.(*types.IntField)
 						if !ok || intField.Value != v {
 							t.Errorf("Expected field %d to be IntField with value %d, got %v", i, v, field)
